==> code/1. letras.ipynb <==
# Import

%pip install lyricsgenius
%pip install pandas

import json
import pandas as pd
import numpy as np
from lyricsgenius import Genius
import re
import time

# Load data

## Data

df = pd.read_csv('/home/keos_angie/Música/music/Data/Data/data_incial_completa.csv').astype({ 'valence_tags': np.float32, 'arousal_tags':np.float32, 'dominance_tags':np.float32})

## SAMPLE

#df = df.head(100)
df

# Letras

genius = Genius("DSIr4PBOWp6v6O9nk9V6IngPxfqJaz0hRv-DR1EfnnJZ7A63ELjwE5KAAujkSm2y")

k = 61022
tam = df["track"].size
Letra = []
Idioma = []
lyrics_state = []
hasLetra = []
Lyrics = []


while (k < tam):
    print("k",k)
    try:
        if k < tam:
            
            data = genius.search_song(title=df["track"][k], artist=df["artist"][k])
            
            if(data != None) :
                data = data.to_dict()
                Idioma = np.concatenate([Idioma, [data["language"]]])
                lyrics_state = np.concatenate([lyrics_state, [data["lyrics_state"]]])
            
                file_name = "Data/Letras/" + str(k+1) + "-" + str(df["track"][k]).replace(" ","-").replace("/","-") + "-" + str(df["artist"][k]).replace(" ","-").replace("/","-") + ".txt"
                file_name_data = "Data/Data/Data_letras/" + str(k+1) + "-NONE-" + str(df["track"][k]).replace(" ","-").replace("/","-") + "-" + str(df["artist"][k]).replace(" ","-").replace("/","-") + ".json"
                Lyrics = np.concatenate([Lyrics, [data["lyrics"]]])
                Letra = np.concatenate([Letra, [file_name]])
                hasLetra = np.concatenate([hasLetra, [1]])
                with open(file_name, "w") as f:
                    json.dump(data["lyrics"], f)
                with open(file_name_data, "w") as f:
                    json.dump(data, f)
                    
            else:
                Idioma = np.concatenate([Idioma, [0]])
                lyrics_state = np.concatenate([lyrics_state, [0]])
                file_name = "Data/Letras/" + str(k+1) + "-NONE-" + str(df["track"][k]).replace(" ","-").replace("/","-") + "-" + str(df["artist"][k]).replace(" ","-").replace("/","-") + ".txt"
                file_name_data = "Data/Data/Data_letras/" + str(k+1) + "-NONE-" + str(df["track"][k]).replace(" ","-").replace("/","-") + "-" + str(df["artist"][k]).replace(" ","-").replace("/","-") + ".json"
                Letra = np.concatenate([Letra, [file_name]])
                Lyrics = np.concatenate([Lyrics, [""]])
                hasLetra = np.concatenate([hasLetra, [0]])
                
                with open(file_name, "w") as f:
                    json.dump("", f)
                with open(file_name_data, "w") as f:
                    json.dump(data, f)
        letras = {
            'Letra': Letra, 
            'Idioma': Idioma, 
            'lyrics_state': lyrics_state, 
            'hasLetra': hasLetra,
            'Lyrics': Lyrics
        }
        letras_ = pd.DataFrame(letras)
        letras_.to_csv('Data/Data/letras/Data_Letras43.csv', header=True, index=False)
        k = k + 1
            
    except:
        time.sleep(100)
    

print(len(Lyrics))
len(Lyrics)

letras = {
    'Letra': Letra, 
    'Idioma': Idioma, 
    'lyrics_state': lyrics_state, 
    'hasLetra': hasLetra,
    'Lyrics': Lyrics
}
df = pd.DataFrame(letras)

letras

# save

df.to_csv('Data/Data/Data_Letras.csv', header=True, index=False)
