==> code/2. Features preproces.ipynb <==
# Import

%pip install matplotlib
import json
import pandas as pd
import numpy as np
import spotipy
from spotipy.oauth2 import SpotifyClientCredentials
import re

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import matplotlib
print(matplotlib.__version__)

# Audio features

## Low features

low1 = pd.read_csv('Data/Data/data_caracteristicas_audio_low_1.csv')
low2 = pd.read_csv('Data/Data/data_caracteristicas_audio_low_2.csv')
low3 = pd.read_csv('Data/Data/data_caracteristicas_audio_low_3.csv')
low4 = pd.read_csv('Data/Data/data_caracteristicas_audio_low_4.csv')
low5 = pd.read_csv('Data/Data/data_caracteristicas_audio_low_5.csv')
low6 = pd.read_csv('Data/Data/data_caracteristicas_audio_low_6.csv')
low7 = pd.read_csv('Data/Data/data_caracteristicas_audio_low_7.csv')
low9 = pd.read_csv('Data/Data/data_caracteristicas_audio_low_9.csv')
low10 = pd.read_csv('Data/Data/data_caracteristicas_audio_low_10.csv')
low11 = pd.read_csv('Data/Data/data_caracteristicas_audio_low_11.csv')
low12 = pd.read_csv('Data/Data/data_caracteristicas_audio_low_12.csv')
low13 = pd.read_csv('Data/Data/data_caracteristicas_audio_low_13.csv')
low14 = pd.read_csv('Data/Data/data_caracteristicas_audio_low_14.csv')
low15 = pd.read_csv('Data/Data/data_caracteristicas_audio_low_15.csv')
low16 = pd.read_csv('Data/Data/data_caracteristicas_audio_low_16.csv')
low17 = pd.read_csv('Data/Data/data_caracteristicas_audio_low_17.csv')
low18 = pd.read_csv('Data/Data/data_caracteristicas_audio_low_18.csv')
low19 = pd.read_csv('Data/Data/data_caracteristicas_audio_low_19.csv')
low20 = pd.read_csv('Data/Data/data_caracteristicas_audio_low_20.csv')
low21 = pd.read_csv('Data/Data/data_caracteristicas_audio_low_21.csv')
low22 = pd.read_csv('Data/Data/data_caracteristicas_audio_low_22.csv')
low23 = pd.read_csv('Data/Data/data_caracteristicas_audio_low_23.csv')
low24 = pd.read_csv('Data/Data/data_caracteristicas_audio_low_24.csv')
low25 = pd.read_csv('Data/Data/data_caracteristicas_audio_low_25.csv')
low26 = pd.read_csv('Data/Data/data_caracteristicas_audio_low_26.csv')


df = pd.concat([low1,low2,low3,low4,low5,low6,low7,low9,low10,low11,low12,
low13,low14,low15,low16,low17,low18,low19,low20,low21,low22,low23,low24,low25,low26], axis=0)


df

df = df.drop_duplicates()
df

df.to_csv('featuresLow_data.csv', header=True, index=False)
