==> code/3. letras_preproces.ipynb <==
# Import

%pip install spacy
%pip install “modin[all]” # Install Modin dependencies and modin’s execution engines
%pip install -U modin 

import json
import pandas as pd
#import modin.pandas as pd
import numpy as np
from lyricsgenius import Genius
import re
import time
import spacy
import re

# Load data

## Data

letras1 = pd.read_csv('Data/Data/letras/Data_Letras1.csv')
letras2 = pd.read_csv('Data/Data/letras/Data_Letras2.csv')
letras3 = pd.read_csv('Data/Data/letras/Data_Letras3.csv')
letras4 = pd.read_csv('Data/Data/letras/Data_Letras4.csv')
letras5 = pd.read_csv('Data/Data/letras/Data_Letras5.csv')
letras6 = pd.read_csv('Data/Data/letras/Data_Letras6.csv')
letras7 = pd.read_csv('Data/Data/letras/Data_Letras7.csv')
letras8 = pd.read_csv('Data/Data/letras/Data_Letras8.csv')
letras9 = pd.read_csv('Data/Data/letras/Data_Letras9.csv')
letras10 = pd.read_csv('Data/Data/letras/Data_Letras10.csv')
letras11 = pd.read_csv('Data/Data/letras/Data_Letras11.csv')
letras12 = pd.read_csv('Data/Data/letras/Data_Letras12.csv')
letras13 = pd.read_csv('Data/Data/letras/Data_Letras13.csv')
letras14 = pd.read_csv('Data/Data/letras/Data_Letras14.csv')
letras15 = pd.read_csv('Data/Data/letras/Data_Letras15.csv')
letras16 = pd.read_csv('Data/Data/letras/Data_Letras16.csv')
letras17 = pd.read_csv('Data/Data/letras/Data_Letras17.csv')
letras18 = pd.read_csv('Data/Data/letras/Data_Letras18.csv')
letras19 = pd.read_csv('Data/Data/letras/Data_Letras19.csv')
letras20 = pd.read_csv('Data/Data/letras/Data_Letras20.csv')
letras21 = pd.read_csv('Data/Data/letras/Data_Letras21.csv')
letras22 = pd.read_csv('Data/Data/letras/Data_Letras22.csv')
letras23 = pd.read_csv('Data/Data/letras/Data_Letras23.csv')
letras24 = pd.read_csv('Data/Data/letras/Data_Letras24.csv')
letras25 = pd.read_csv('Data/Data/letras/Data_Letras25.csv')
letras26 = pd.read_csv('Data/Data/letras/Data_Letras26.csv')
letras27 = pd.read_csv('Data/Data/letras/Data_Letras27.csv')
letras28 = pd.read_csv('Data/Data/letras/Data_Letras28.csv')
letras29 = pd.read_csv('Data/Data/letras/Data_Letras29.csv')
letras30 = pd.read_csv('Data/Data/letras/Data_Letras30.csv')
letras31 = pd.read_csv('Data/Data/letras/Data_Letras31.csv')
letras32 = pd.read_csv('Data/Data/letras/Data_Letras32.csv')
letras33 = pd.read_csv('Data/Data/letras/Data_Letras33.csv')
letras34 = pd.read_csv('Data/Data/letras/Data_Letras34.csv')
letras35 = pd.read_csv('Data/Data/letras/Data_Letras35.csv')
letras36 = pd.read_csv('Data/Data/letras/Data_Letras36.csv')
letras37 = pd.read_csv('Data/Data/letras/Data_Letras37.csv')
letras38 = pd.read_csv('Data/Data/letras/Data_Letras38.csv')
letras39 = pd.read_csv('Data/Data/letras/Data_Letras39.csv')
letras40 = pd.read_csv('Data/Data/letras/Data_Letras40.csv')
letras41 = pd.read_csv('Data/Data/letras/Data_Letras41.csv')
letras42 = pd.read_csv('Data/Data/letras/Data_Letras42.csv')
letras43 = pd.read_csv('Data/Data/letras/Data_Letras43.csv')
df = pd.concat([letras1, 
letras2, 
letras3, letras4, letras5, letras6, 
letras7, letras8, letras9, letras10, letras11,
letras12, letras13, letras14, letras15, letras16,
letras17, letras18, letras19, letras20, letras21,
letras22, letras23, letras24, letras25, letras26,
letras27, letras28, letras29, letras30, letras31, letras32,
letras33, letras34, letras35, letras36, letras37, letras38,
letras39, letras40, letras41, letras42, letras43], axis=0)


df = df.drop_duplicates()
df

df.to_csv('letras_data.csv', header=True, index=False)

nlp = spacy.blank("en")
url_pattern = re.compile(r"https?://[^,; ]+")
char_pattern = re.compile(r"[^a-z ]")
spaces_pattern = re.compile(r"\s+")
parts_pattern = re.compile(r"\[[^]]*\]")
noIntro_pattern = re.compile(r"^.*Lyrics")
noFinal_pattern = re.compile(r"\d*embed")

def preprocess(text: str) -> str:
	try:
		noIntro = re.sub(noIntro_pattern, " ", text)
		lower_text = str(noIntro).lower()
		no_parts = re.sub(parts_pattern, " ", lower_text)
		no_double_spaces = re.sub(spaces_pattern, " ", no_parts)
		noFinal = re.sub(noFinal_pattern, "", no_double_spaces)
		no_special = re.sub(char_pattern, "", noFinal)

		doc = nlp(no_special)
		valid_text = " ".join(
			map(
				lambda token: token.text,
				filter(lambda token: (
					not token.is_stop and
					len(token.text) < 24 and
					len(token.text) > 2
				), doc)
			)
		)
		return valid_text.strip()
	except:
		return None

print(preprocess(df.Lyrics.iloc[1]))

df["Lyrics_proces"] = df.Lyrics.apply(preprocess)

corpus_prep = df.Lyrics_proces

corpus_prep

df

df.to_csv('letras_preproces.csv', header=True, index=False)
