==> code/4. data_unificación.ipynb <==
# Import

%pip install matplotlib
%pip install fuzzywuzzy
import json
import pandas as pd
import numpy as np
import spotipy
from spotipy.oauth2 import SpotifyClientCredentials
import re

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import matplotlib
print(matplotlib.__version__)
from fuzzywuzzy import fuzz

# Data

letras = pd.read_csv('letras_preproces.csv')
audio = pd.read_csv('featuresLow_data.csv')

# LETRAS SONG ID
songId = []
for k, song in enumerate(letras["Letra"]):
	aux = song.split("-", maxsplit=1)[-1]
	aux = aux.strip().lower()
	aux = aux.replace('none-','').replace("'","").replace("...","").replace("the-","").replace("(feat.-in-flames)-","").replace("---remastered","")
	songId.append(aux)
letras["songId"] = songId

# AUDIO SONG ID
songId = []
for k, song in enumerate(audio["track_uri"]):
	
	aux = str(audio["track_name"][k]).replace(" ","-").replace("/","-").replace("'","").lower() \
    + "-" + str(audio["artist_name"][k]).replace(" ","-").replace("/","-").replace("'","").lower() \
    + ".txt"

	aux = aux.replace("---main","").replace("sehen-","").replace("...","").replace("the-","").replace("(feat.-in-flames)-","").replace("---remastered","")
	songId.append(aux)
audio["songId"] = songId

# Utilizar fuzzywuzzy para encontrar coincidencias cercanas
def fuzzy_merge(df1, df2, key1, key2, threshold=90, limit=2):
    """
    Realiza una fusión entre dos dataframes utilizando la librería fuzzywuzzy para encontrar las coincidencias más cercanas.
    """
    s = df2[key2].tolist()

    m = df1[key1].apply(lambda x: pd.Series([fuzz.token_set_ratio(x, i) for i in s if isinstance(i, str)]).nlargest(limit))

    df1['matches'] = m.apply(lambda x: ', '.join(s[x.index[x >= threshold]].tolist()), axis=1)

    return pd.merge(df1, df2, left_on='matches', right_on=key2, how='left')



# Realizar la fusión utilizando los identificadores únicos
#merged_df = fuzzy_merge(letras, audio, 'songId', 'songId')


df = pd.merge(letras, audio, on='songId', how='outer')
df


not_matched = df["track_uri"].isna()
#not_matched = df_merged.isna().any(axis=1)

# Mostrar los registros que no hicieron match
df[not_matched]


print(audio.loc[89]["songId"])
print(letras.loc[89]["songId"])

df.dropna(subset=['liveness'], inplace=True)
df.dropna(subset=['Letra'], inplace=True)

df

df.to_csv('Data_unificada.csv', header=True, index=False)
