==> code/6. Data EDA.ipynb <==
# Import

# Importaciones necesarias
%pip  install seaborn
import pandas as pd
import numpy as np
from numpy import nan
from numpy.random import randn
from numpy import percentile
import seaborn as sns
import statistics
from statistics import mode
import json
import pandas as pd
import numpy as np
import re
%pip install wordcloud
from wordcloud import WordCloud
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

from sklearn.preprocessing import StandardScaler
import matplotlib.pyplot as plt
from numpy import asarray
from sklearn.preprocessing import OneHotEncoder
import matplotlib.pyplot as plt
import matplotlib

df = pd.read_csv('/content/data_incial_completa.csv')

df.head(10)

from google.colab import drive
drive.mount('/content/drive')

df.shape

df.info()

## EDA

numerical_columns = df.select_dtypes(include=np.number).columns
list(numerical_columns)

categorical_columns = df.select_dtypes(exclude=np.number).columns
categorical_columns

original_data = np.array(df['arousal_tags']).reshape(-1,1)

std = StandardScaler()
normalized_data = std.fit_transform(original_data)

fig, ax = plt.subplots(1, 2, figsize=(15, 6))
sns.histplot(original_data, ax=ax[0], kde=True, legend=False)
ax[0].set_title("Original Data")
sns.histplot(normalized_data, ax=ax[1], kde=True, legend=False)
ax[1].set_title("Normalized data")
plt.show()

df['arousal_tags'] = normalized_data

original_data = np.array(df['valence_tags']).reshape(-1,1)

std = StandardScaler()
normalized_data = std.fit_transform(original_data)

fig, ax = plt.subplots(1, 2, figsize=(15, 6))
sns.histplot(original_data, ax=ax[0], kde=True, legend=False)
ax[0].set_title("Original Data")
sns.histplot(normalized_data, ax=ax[1], kde=True, legend=False)
ax[1].set_title("Normalized data")
plt.show()

df['valence_tags'] = normalized_data

## Relación de data social vs cuadrante etiqueta

# crear la figura con un tamaño adecuado
fig, ax = plt.subplots(figsize=(4, 3), dpi=100);

#matplotlib.rcParams['font.family'] = 'Arial'
matplotlib.rcParams['font.size'] = 12;

# trazar un scatter plot con separación por especie
sns.scatterplot(x='valence_tags', y='arousal_tags', hue='cuadrante', data=df, ax=ax, palette='viridis');

# agregar etiquetas y título
ax.set_xlabel('Valence')
ax.set_ylabel('Arousal')
ax.legend(title='Quadrant', title_fontsize=12, fontsize=10)  # Cambiar nombre de leyenda
#ax.set_title('Relación entre Valencia y Arousal según cuadrante \n para datos socialmente etiquetados')

# agregar líneas de referencia
ax.axhline(y=0, color='black', linestyle='--')
ax.axvline(x=0, color='black', linestyle='--')

# eliminar marcas de rejilla
ax.grid(False)

# mostrar el gráfico
plt.show()

# guardar la figura en formato PNG
plt.savefig('mi_grafica.png', dpi=300, bbox_inches='tight')


### relación por cuadrante

import seaborn as sns
import matplotlib.pyplot as plt
import seaborn as sns

#matplotlib.rcParams['font.family'] = 'Arial'
matplotlib.rcParams['font.size'] = 12

df_1 = df[df['cuadrante']==1]
df_2 = df[df['cuadrante']==2]
df_3 = df[df['cuadrante']==3]
df_4 = df[df['cuadrante']==4]
# Crear una figura y dos ejes con la función subplots de Matplotlib
fig, axes = plt.subplots(nrows=2, ncols=2, figsize=(10, 8), dpi=300)

# trazar un scatter plot con separación por especie
sns.scatterplot(x='valence_tags', y='arousal_tags', hue = 'cuadrante',  data=df_2, legend=False, ax=axes[0][0], palette='viridis')
axes[0][0].axhline(y=0, color='black', linestyle='--')
axes[0][0].axvline(x=0, color='black', linestyle='--')
# agregar etiquetas y título
axes[0][0].set_xlabel('Valence', fontsize=12)
axes[0][0].set_ylabel('Arousal', fontsize=12)
axes[0][0].set_title('A. Quadrant 2', fontsize=14)


# Graficar la segunda gráfica de barras en el segundo eje
sns.scatterplot(x='valence_tags', y='arousal_tags', hue = 'cuadrante',  data=df_1, legend=False, ax=axes[0][1], palette='magma')
axes[0][1].axhline(y=0, color='black', linestyle='--')
axes[0][1].axvline(x=0, color='black', linestyle='--')
axes[0][1].set_xlabel('Valence', fontsize=12)
axes[0][1].set_ylabel('Arousal', fontsize=12)
axes[0][1].set_title('B. Quadrant 1', fontsize=14)

# Graficar la segunda gráfica de barras en el segundo eje
sns.scatterplot(x='valence_tags', y='arousal_tags', hue = 'cuadrante',  data=df_3, legend=False, ax=axes[1][0], palette='coolwarm')
axes[1][0].axhline(y=0, color='black', linestyle='--')
axes[1][0].axvline(x=0, color='black', linestyle='--')
axes[1][0].set_xlabel('Valence', fontsize=12)
axes[1][0].set_ylabel('Arousal', fontsize=12)
axes[1][0].set_title('C. Quadrant 3', fontsize=14)

# Graficar la segunda gráfica de barras en el segundo eje
sns.scatterplot(x='valence_tags', y='arousal_tags', hue = 'cuadrante',  data=df_4, legend=False, ax=axes[1][1], palette='plasma')
axes[1][1].axhline(y=0, color='black', linestyle='--')
axes[1][1].axvline(x=0, color='black', linestyle='--')
axes[1][1].set_xlabel('Valence', fontsize=12)
axes[1][1].set_ylabel('Arousal', fontsize=12)
axes[1][1].set_title('D. Quadrant 4', fontsize=14)

# Agregar título al conjunto de gráficas
#fig.suptitle('Distribución de los datos según valores de valence y aurosal por cuadrante', fontsize=14)

fig.subplots_adjust(hspace=0.5, wspace=0.2)

# Calcular el promedio por cuadrante
df_promedios = df.groupby(['cuadrante']).mean()
#df_promedios = df_promedios["valence_tags"]
axes[0][0].text(0.05, 0.95, f"Promedio:\nValence={df_promedios['valence_tags'][2]:.2f}\nArousal={df_promedios['arousal_tags'][1]:.2f}", transform=axes[0][0].transAxes, va='top', fontsize=12, bbox=dict(boxstyle='round', facecolor='white', alpha=0.8))
axes[0][1].text(0.05, 0.95, f"Promedio:\nValence={df_promedios['valence_tags'][1]:.2f}\nArousal={df_promedios['arousal_tags'][2]:.2f}", transform=axes[0][1].transAxes, va='top', fontsize=12, bbox=dict(boxstyle='round', facecolor='white', alpha=0.8))
axes[1][0].text(0.05, 0.95, f"Promedio:\nValence={df_promedios['valence_tags'][3]:.2f}\nArousal={df_promedios['arousal_tags'][3]:.2f}", transform=axes[1][0].transAxes, va='top', fontsize=12, bbox=dict(boxstyle='round', facecolor='white', alpha=0.8))
axes[1][1].text(0.05, 0.95, f"Promedio:\nValence={df_promedios['valence_tags'][4]:.2f}\nArousal={df_promedios['arousal_tags'][4]:.2f}", transform=axes[1][1].transAxes, va='top', fontsize=12, bbox=dict(boxstyle='round', facecolor='white', alpha=0.8))

# Mostrar la figura
plt.show()

import seaborn as sns
import matplotlib.pyplot as plt
import seaborn as sns

#matplotlib.rcParams['font.family'] = 'Arial'
matplotlib.rcParams['font.size'] = 12

df_1 = df[df['cuadrante']==1]
df_2 = df[df['cuadrante']==2]
df_3 = df[df['cuadrante']==3]
df_4 = df[df['cuadrante']==4]
# Crear una figura y dos ejes con la función subplots de Matplotlib
fig, axes = plt.subplots(nrows=2, ncols=2, figsize=(10, 8), dpi=300)

# trazar un scatter plot con separación por especie
sns.scatterplot(x='valence_tags', y='arousal_tags', hue = 'cuadrante',  data=df_2, legend=False, ax=axes[0][0], palette='viridis')
axes[0][0].axhline(y=0, color='black', linestyle='--')
axes[0][0].axvline(x=0, color='black', linestyle='--')
# agregar etiquetas y título
axes[0][0].set_xlabel('Valence', fontsize=12)
axes[0][0].set_ylabel('Arousal', fontsize=12)
axes[0][0].set_title('A. Quadrant 2', fontsize=14)


# Graficar la segunda gráfica de barras en el segundo eje
sns.scatterplot(x='valence_tags', y='arousal_tags', hue = 'cuadrante',  data=df_1, legend=False, ax=axes[0][1], palette='magma')
axes[0][1].axhline(y=0, color='black', linestyle='--')
axes[0][1].axvline(x=0, color='black', linestyle='--')
axes[0][1].set_xlabel('Valence', fontsize=12)
axes[0][1].set_ylabel('Arousal', fontsize=12)
axes[0][1].set_title('B. Quadrant 1', fontsize=14)

# Graficar la segunda gráfica de barras en el segundo eje
sns.scatterplot(x='valence_tags', y='arousal_tags', hue = 'cuadrante',  data=df_3, legend=False, ax=axes[1][0], palette='coolwarm')
axes[1][0].axhline(y=0, color='black', linestyle='--')
axes[1][0].axvline(x=0, color='black', linestyle='--')
axes[1][0].set_xlabel('Valence', fontsize=12)
axes[1][0].set_ylabel('Arousal', fontsize=12)
axes[1][0].set_title('C. Quadrant 3', fontsize=14)

# Graficar la segunda gráfica de barras en el segundo eje
sns.scatterplot(x='valence_tags', y='arousal_tags', hue = 'cuadrante',  data=df_4, legend=False, ax=axes[1][1], palette='plasma')
axes[1][1].axhline(y=0, color='black', linestyle='--')
axes[1][1].axvline(x=0, color='black', linestyle='--')
axes[1][1].set_xlabel('Valence', fontsize=12)
axes[1][1].set_ylabel('Arousal', fontsize=12)
axes[1][1].set_title('D. Quadrant 4', fontsize=14)

# Agregar título al conjunto de gráficas
#fig.suptitle('Distribución de los datos según valores de valence y aurosal por cuadrante', fontsize=14)

fig.subplots_adjust(hspace=0.5, wspace=0.2)

# Calcular el promedio por cuadrante
df_promedios = df.groupby(['cuadReal']).mean()
#df_promedios = df_promedios["valence_tags"]
axes[0][0].text(0.05, 0.95, f"Promedio:\nTempo={df_promedios['tempo'][2]:.2f}", transform=axes[0][0].transAxes, va='top', fontsize=12, bbox=dict(boxstyle='round', facecolor='white', alpha=0.8))
axes[0][1].text(0.05, 0.95, f"Promedio:\nTempo={df_promedios['tempo'][1]:.2f}", transform=axes[0][1].transAxes, va='top', fontsize=12, bbox=dict(boxstyle='round', facecolor='white', alpha=0.8))
axes[1][0].text(0.05, 0.95, f"Promedio:\nTempo={df_promedios['tempo'][3]:.2f}", transform=axes[1][0].transAxes, va='top', fontsize=12, bbox=dict(boxstyle='round', facecolor='white', alpha=0.8))
axes[1][1].text(0.05, 0.95, f"Promedio:\nTempo={df_promedios['tempo'][4]:.2f}", transform=axes[1][1].transAxes, va='top', fontsize=12, bbox=dict(boxstyle='round', facecolor='white', alpha=0.8))

# Mostrar la figura
plt.show()

import seaborn as sns
import matplotlib.pyplot as plt
import seaborn as sns

#matplotlib.rcParams['font.family'] = 'Arial'
matplotlib.rcParams['font.size'] = 12

df_1 = df[df['cuadrante']==1]
df_2 = df[df['cuadrante']==2]
df_3 = df[df['cuadrante']==3]
df_4 = df[df['cuadrante']==4]
# Crear una figura y dos ejes con la función subplots de Matplotlib
fig, axes = plt.subplots(nrows=2, ncols=2, figsize=(10, 8), dpi=300)

# trazar un scatter plot con separación por especie
sns.scatterplot(x='valence_tags', y='arousal_tags', hue = 'cuadrante',  data=df_2, legend=False, ax=axes[0][0], palette='viridis')
axes[0][0].axhline(y=0, color='black', linestyle='--')
axes[0][0].axvline(x=0, color='black', linestyle='--')
# agregar etiquetas y título
axes[0][0].set_xlabel('Valence', fontsize=12)
axes[0][0].set_ylabel('Arousal', fontsize=12)
axes[0][0].set_title('A. Quadrant 2', fontsize=14)


# Graficar la segunda gráfica de barras en el segundo eje
sns.scatterplot(x='valence_tags', y='arousal_tags', hue = 'cuadrante',  data=df_1, legend=False, ax=axes[0][1], palette='magma')
axes[0][1].axhline(y=0, color='black', linestyle='--')
axes[0][1].axvline(x=0, color='black', linestyle='--')
axes[0][1].set_xlabel('Valence', fontsize=12)
axes[0][1].set_ylabel('Arousal', fontsize=12)
axes[0][1].set_title('B. Quadrant 1', fontsize=14)

# Graficar la segunda gráfica de barras en el segundo eje
sns.scatterplot(x='valence_tags', y='arousal_tags', hue = 'cuadrante',  data=df_3, legend=False, ax=axes[1][0], palette='coolwarm')
axes[1][0].axhline(y=0, color='black', linestyle='--')
axes[1][0].axvline(x=0, color='black', linestyle='--')
axes[1][0].set_xlabel('Valence', fontsize=12)
axes[1][0].set_ylabel('Arousal', fontsize=12)
axes[1][0].set_title('C. Quadrant 3', fontsize=14)

# Graficar la segunda gráfica de barras en el segundo eje
sns.scatterplot(x='valence_tags', y='arousal_tags', hue = 'cuadrante',  data=df_4, legend=False, ax=axes[1][1], palette='plasma')
axes[1][1].axhline(y=0, color='black', linestyle='--')
axes[1][1].axvline(x=0, color='black', linestyle='--')
axes[1][1].set_xlabel('Valence', fontsize=12)
axes[1][1].set_ylabel('Arousal', fontsize=12)
axes[1][1].set_title('D. Quadrant 4', fontsize=14)

# Agregar título al conjunto de gráficas
#fig.suptitle('Distribución de los datos según valores de valence y aurosal por cuadrante', fontsize=14)

fig.subplots_adjust(hspace=0.5, wspace=0.2)

# Calcular el promedio por cuadrante
df_promedios = df.groupby(['cuadReal']).mean()
#df_promedios = df_promedios["valence_tags"]
axes[0][0].text(0.05, 0.95, f"Promedio:\n# Emotion Tags={df_promedios['number_of_emotion_tags'][2]:.2f}", transform=axes[0][0].transAxes, va='top', fontsize=12, bbox=dict(boxstyle='round', facecolor='white', alpha=0.8))
axes[0][1].text(0.05, 0.95, f"Promedio:\n# Emotion Tags={df_promedios['number_of_emotion_tags'][1]:.2f}", transform=axes[0][1].transAxes, va='top', fontsize=12, bbox=dict(boxstyle='round', facecolor='white', alpha=0.8))
axes[1][0].text(0.05, 0.95, f"Promedio:\n# Emotion Tags={df_promedios['number_of_emotion_tags'][3]:.2f}", transform=axes[1][0].transAxes, va='top', fontsize=12, bbox=dict(boxstyle='round', facecolor='white', alpha=0.8))
axes[1][1].text(0.05, 0.95, f"Promedio:\n# Emotion Tags={df_promedios['number_of_emotion_tags'][4]:.2f}", transform=axes[1][1].transAxes, va='top', fontsize=12, bbox=dict(boxstyle='round', facecolor='white', alpha=0.8))

# Mostrar la figura
plt.show()

df_promedios[['valence_tags','arousal_tags']]

df_promedios = df.groupby(['cuadReal']).mean()
df_promedios[['valence_tags','arousal_tags','number_of_emotion_tags','tempo']]

df_promedios = df.groupby(['cuadReal']).mean()
df_promedios[['danceability', 'energy', 'key',
       'loudness', 'mode', 'speechiness', 'acousticness', 'instrumentalness',
       'liveness', 'valence', 'tempo', 'duration_ms', 'time_signature',
       'number_of_emotion_tags', 'valence_tags', 'arousal_tags',
       'dominance_tags']]

df['cuadReal'].value_counts()


df['cuadrante'].value_counts()

df_promedios = df.groupby(['key']).agg(lambda x: x.mode())
df_promedios[['cuadReal','cuadrante']]

df_grouped = df.groupby(['cuadReal', 'key']).size().reset_index(name='counts')
df_grouped

## crear la figura con un tamaño adecuado
#fig, ax = plt.subplots(figsize=(8, 6), dpi=300);

##matplotlib.rcParams['font.family'] = 'Arial'
#matplotlib.rcParams['font.size'] = 12;

## trazar un scatter plot con separación por especie
##sns.scatterplot(x='tempo', y='duration_ms', data=df, ax=ax, palette='viridis');
#sns.pairplot(df)
## agregar etiquetas y título
## eliminar marcas de rejilla
#ax.grid(False)

# mostrar el gráfico
#plt.show()


## Cuadrantes reales spotify

# crear la figura con un tamaño adecuado
fig, ax = plt.subplots(figsize=(4, 3), dpi=100);

#matplotlib.rcParams['font.family'] = 'Arial'
matplotlib.rcParams['font.size'] = 12

# trazar un scatter plot con separación por especie
sns.scatterplot(x='valence', y='energy', hue = df.cuadReal,  data=df,  palette='viridis')

# agregar etiquetas y título
ax.set_xlabel('Valence')
ax.set_ylabel('Arousal')

#ax.set_title('Relación entre Valencia y Arousal según cuadrante \n para datos extraidos de la RRNN de spotify')

# agregar líneas de referencia
ax.axhline(y=0.5, color='black', linestyle='--')
ax.axvline(x=0.5, color='black', linestyle='--')
ax.legend(title='Quadrant', title_fontsize=12, fontsize=10)  # Cambiar nombre de leyenda
# eliminar marcas de rejilla
ax.grid(False)

# mostrar el gráfico
plt.show()

# guardar la figura en formato PNG
plt.savefig('mi_grafica.png', dpi=300, bbox_inches='tight')

## Relación con genero

df.genre.value_counts()

# crear la figura con un tamaño adecuado
fig, ax = plt.subplots(figsize=(8, 6), dpi=300);

#matplotlib.rcParams['font.family'] = 'Arial'
matplotlib.rcParams['font.size'] = 12

# trazar un scatter plot con separación por especie
sns.scatterplot(x='valence', y='energy', hue = df["genre"],alpha=0.5,  data=df, legend=False, palette='viridis')

# agregar etiquetas y título
ax.set_xlabel('Valence')
ax.set_ylabel('Arousal')
#ax.set_title('Relación entre Valencia y Arousal según cuadrante \n para datos extraidos de la RRNN de spotify')

# agregar líneas de referencia
ax.axhline(y=0.5, color='black', linestyle='--')
ax.axvline(x=0.5, color='black', linestyle='--')

# eliminar marcas de rejilla
ax.grid(False)

# mostrar el gráfico
plt.show()


# WC

cuadrante1 = {}
cuadrante2 = {}
cuadrante3 = {}
cuadrante4 = {}

for index, row in df.iterrows():
    if row['cuadrante'] == 1:
        words = row['seeds'].split()
        for word in words:
            if word in cuadrante1:
                cuadrante1[word] += 1
            else:
                cuadrante1[word] = 1
    elif row['cuadrante'] == 2:
        words = row['seeds'].split()
        for word in words:
            if word in cuadrante2:
                cuadrante2[word] += 1
            else:
                cuadrante2[word] = 1
    elif row['cuadrante'] == 3:
        words = row['seeds'].split()
        for word in words:
            if word in cuadrante3:
                cuadrante3[word] += 1
            else:
                cuadrante3[word] = 1
    elif row['cuadrante'] == 4:
        words = row['seeds'].split()
        for word in words:
            if word in cuadrante4:
                cuadrante4[word] += 1
            else:
                cuadrante4[word] = 1


wordcloud1 = WordCloud(width=1500, height=1000, background_color='white', colormap='viridis').generate_from_frequencies(cuadrante1)
wordcloud2 = WordCloud(width=1500, height=1000, background_color='white', colormap='viridis').generate_from_frequencies(cuadrante2)
wordcloud3 = WordCloud(width=1500, height=1000, background_color='white', colormap='viridis').generate_from_frequencies(cuadrante3)
wordcloud4 = WordCloud(width=1500, height=1000, background_color='white', colormap='viridis').generate_from_frequencies(cuadrante4)


fig, axs = plt.subplots(2, 2,figsize=(8, 6), dpi=300)
matplotlib.rcParams['font.size'] = 12

axs[0, 0].imshow(wordcloud2)
axs[0, 0].set_title('Quadrant 2')

axs[0, 1].imshow(wordcloud1)
axs[0, 1].set_title('Quadrant 1')

axs[1, 0].imshow(wordcloud3)
axs[1, 0].set_title('Quadrant 3')

axs[1, 1].imshow(wordcloud4)
axs[1, 1].set_title('Quadrant 4')

for ax in axs.flat:
    ax.axis('off')

plt.show()


df.groupby(['cuadReal']).mean()

df['key'].value_counts()

# letras


df = df[df["emocion_mas_comun"] > 0]
df = df[df["hasLetra"] >0]

# crear la figura con un tamaño adecuado
fig, ax = plt.subplots(figsize=(8, 6), dpi=300)

#matplotlib.rcParams['font.family'] = 'Arial'
matplotlib.rcParams['font.size'] = 12

# trazar un scatter plot con separación por especie
sns.scatterplot(x='valence', y='energy', hue = "emocion_mas_comun",  data=df,  palette='viridis')

# agregar etiquetas y título
ax.set_xlabel('Valence')
ax.set_ylabel('Arousal')
#ax.set_title('Relación entre Valencia y Arousal según cuadrante \n para datos extraidos de la RRNN de spotify')

# agregar líneas de referencia
ax.axhline(y=0.5, color='black', linestyle='--')
ax.axvline(x=0.5, color='black', linestyle='--')
ax.legend(title='Quadrant', title_fontsize=12, fontsize=10)

# eliminar marcas de rejilla
ax.grid(False)

# mostrar el gráfico
plt.show()

# guardar la figura en formato PNG
plt.savefig('mi_grafica.png', dpi=300, bbox_inches='tight')

#spotify vs etiquetas
diference_cuad_emocion_mas_comun = 0
normi = 0

for k, i in enumerate(df["cuadrante"]):
  if(df["cuadReal"].iloc[k] == df["cuadrante"].iloc[k]):
    diference_cuad_emocion_mas_comun += 1
  else:
    normi +=1
diference_cuad_emocion_mas_comun , normi, len(df["cuadReal"])

#letra vs etiquetas
diference_cuad_emocion_mas_comun = 0
normi = 0

for k, i in enumerate(df["cuadrante"]):
  if(df["cuadrante"].iloc[k] == df["emocion_mas_comun"].iloc[k]):
    diference_cuad_emocion_mas_comun += 1
  else:
    normi +=1
diference_cuad_emocion_mas_comun , normi, len(df["cuadrante"])

#letra vs spotify
diference_cuad_emocion_mas_comun = 0
normi = 0

for k, i in enumerate(df["cuadReal"]):
  if(df["cuadReal"].iloc[k] == df["emocion_mas_comun"].iloc[k]):
    diference_cuad_emocion_mas_comun += 1
  else:
    normi +=1

diference_cuad_emocion_mas_comun , normi, len(df["cuadReal"])


# crear la figura con un tamaño adecuado
fig, ax = plt.subplots(figsize=(8, 6), dpi=300)

#matplotlib.rcParams['font.family'] = 'Arial'
matplotlib.rcParams['font.size'] = 12


# Calcular las proporciones
total = len(df["cuadrante"])
prop_dif_cuad_emocion_mas_comun = diference_cuad_emocion_mas_comun / total
prop_normi = normi / total
print(total)
print(prop_normi)
print(prop_dif_cuad_emocion_mas_comun)
print(prop_dif_cuad_emocion_mas_comun+prop_normi)

df_prop = {'prop': [prop_dif_cuad_emocion_mas_comun, prop_normi], 'variable': ["Songs with \nquadrant matching", "Songs without \nquadrant matching"]}

# trazar un scatter plot con separación por especie
sns.barplot(x='variable', y= 'prop',  data=df_prop,  palette='Spectral')
# agregar etiquetas de anotación en cada barra
for i, v in enumerate(df_prop['prop']):
    ax.annotate(str(round(v*100,2)) + "%", xy=(i, v), ha='center', va='bottom')

# agregar etiquetas y título
ax.set_ylabel('Song ratio')
#ax.set_ylabel('Arousal')
#ax.set_title('Relación entre Valencia y Arousal según cuadrante \n para datos extraidos de la RRNN de spotify')

# agregar líneas de referencia
#ax.axhline(y=0.5, color='black', linestyle='--')
#ax.axvline(x=0.5, color='black', linestyle='--')
#ax.legend(title='Quadrant', title_fontsize=12, fontsize=10)

# eliminar marcas de rejilla
ax.grid(False)

# mostrar el gráfico
plt.show()

# guardar la figura en formato PNG
#plt.savefig('mi_grafica.png', dpi=300, bbox_inches='tight')

df.to_csv('Data_completa_enrriquecida.csv', header=True, index=False)
