==> code/8_letras_redNeuronal.ipynb <==
# Import

import sys

%pip install joblib==1.2.0
%pip install  pycaret
%pip install scikit-learn

%pip install seaborn

%pip install pytorch-lightning

%pip install plotly

%pip install nltk

%pip install unidecode




import pandas as pd
import numpy as np
import statistics
import json
import torch
import pytorch_lightning as pl
import re, nltk
import torch
import joblib
import seaborn as sns
import matplotlib.pyplot as plt
import matplotlib
import plotly.express as px
import plotly.graph_objects as go

from pycaret.classification import  *

from statistics import mode
from unidecode import unidecode

from nltk.corpus import stopwords
from nltk.tokenize import word_tokenize

from torch import Tensor

from torch.nn import Module, Linear, Dropout, Sequential
from torch.nn.functional import cross_entropy, relu, sigmoid, binary_cross_entropy

from torch.optim import Adam, Optimizer
from torch.utils.data import Dataset, DataLoader, random_split

from pytorch_lightning import LightningModule, Trainer
from pytorch_lightning.loggers import CSVLogger
from pytorch_lightning.callbacks.early_stopping import EarlyStopping


from typing import Tuple, List, Callable

from numpy import nan, percentile, asarray
from numpy.typing import ArrayLike
from numpy.random import randn


from sklearn.preprocessing import OneHotEncoder, StandardScaler
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.decomposition import TruncatedSVD



# Carga de Datos

df = pd.read_csv('Data_completa_enrriquecida.csv',  index_col=0)
df.head()

## preprocesamiento

df = df[['Idioma', 'lyrics_state', 'hasLetra', 'Lyrics', 'Lyrics_proces',
       'songId', 'track_uri', 'emocion_mas_comun', 'cuadrante',
       'cuadReal']]

df.Idioma.value_counts()

df = df[df["Idioma"]=='en']

corpus = df.Lyrics_proces

df.shape

# Extracción de Características

corpus_df = df[['Lyrics', 'cuadReal']]
corpus_df = corpus_df.drop_duplicates()
corpus_df = corpus_df.reset_index()
corpus_df = corpus_df[['Lyrics', 'cuadReal']]


s = setup(corpus_df, 
	target = 'cuadReal', 
	session_id = 123, 
	text_features_method= 'tf-idf', 
	text_features = ['Lyrics'], 
	ignore_features = ['Letra'],
	fix_imbalance = True
)

## Tfidf

vectorizer = TfidfVectorizer(
    sublinear_tf=True, min_df=70, max_df=0.6
    )

tfidf_matrix = vectorizer.fit_transform(corpus)

joblib.dump(tfidf_matrix, "vect_tfidf.joblib")

tfidf_matrix = joblib.load("vect_tfidf.joblib") 

## LSA

# Aplicar LSA con TruncatedSVD
lsa_model = TruncatedSVD(n_components=5)  # Especifica el número de componentes deseados
lsa_matrix = lsa_model.fit_transform(tfidf_matrix)

joblib.dump(lsa_matrix, "lsa_matrix.joblib")

lsa_matrix = joblib.load("lsa_matrix.joblib") 

# Dividir datos

class PersonalityDataset(Dataset):
    def __init__(self, features: ArrayLike, labels: ArrayLike):
        self.features = torch.tensor(features, dtype=torch.float32)
        self.labels = torch.tensor(labels, dtype=torch.int64)

    def __len__(self) -> int:
        return self.features.shape[0]

    def __getitem__(self, idx: int) -> Tuple[Tensor, Tensor]:
        return self.features[idx], self.labels[idx]

features = vect.transform(corpus).toarray()
labels = df.cuadReal.to_numpy()

labels.shape

features.shape

ds = PersonalityDataset(
        features = features, labels=labels
        ) 
train_ds, val_ds, test_ds = random_split(ds, [0.6, 0.2, 0.2])

train_dl = DataLoader(train_ds, batch_size=1024) 
val_dl = DataLoader(val_ds, batch_size=512) 
test_dl = DataLoader(test_ds, batch_size=512) 

(train_ds.dataset.features).shape, train_ds.dataset.labels.shape

# Modelamiento

## Random Forest

from sklearn.ensemble import RandomForestClassifier

model = RandomForestClassifier(max_depth=15, random_state=0)

model.fit(train_ds.dataset.features, train_ds.dataset.labels)
predicted = model.predict(train_ds.dataset.features)

print(f'El score del modelo  es: {model.score(train_ds.dataset.features, train_ds.dataset.labels)}')

## árboles de decisión

from sklearn.tree import DecisionTreeClassifier
from sklearn import tree

model = DecisionTreeClassifier(max_depth=5)

clf = model.fit(train_ds.dataset.features, train_ds.dataset.labels) 
predicted = model.predict(train_ds.dataset.features)

print(f'El score del modelo  es: {model.score(train_ds.dataset.features, train_ds.dataset.labels)}')

## KNN

from sklearn.neighbors import KNeighborsClassifier

model = KNeighborsClassifier(n_neighbors=10)

model.fit(train_ds.dataset.features, train_ds.dataset.labels)
predicted = model.predict(train_ds.dataset.features)

print(f'El score del modelo  es: {model.score(train_ds.dataset.features, train_ds.dataset.labels)}')

## regresión logística

from sklearn.linear_model import LogisticRegression

model = LogisticRegression(solver='lbfgs', max_iter=10)

model.fit(train_ds.dataset.features, train_ds.dataset.labels)
predicted = model.predict(train_ds.dataset.features)

print(f'El score del modelo es: {model.score(train_ds.dataset.features, train_ds.dataset.labels)}')

balancear -> smote (data aumentation)
revisar las proporciones de los cuadrantes
usar modelo preentrenado
probar random forest

Batch = Tuple[Tensor, Tensor]

class Dense(Module):
    def __init__(
        self,
        in_features: int,
        out_features: int,
        activation: Callable,
        dropout: float,
        ):
        super(Dense, self).__init__()
        self.layer = Linear(in_features=in_features, out_features=out_features)
        self.dropout = Dropout(p=dropout)
        self.activation = activation
    
    def forward(self, x: Tensor) -> Tensor:
        return self.activation(self.dropout(self.layer(x)))

class MLP(LightningModule):
    def __init__(
            self, n_units: List[int], activations: List[Callable],
            dropout: float
            ):
        super(MLP, self).__init__()
        self.layers = Sequential(
            *(
                Dense(
                    in_features=ins,
                    out_features=outs,
                    activation=act,
                    dropout=dropout
                    )
                for ins, outs, act in zip(
                    n_units,
                    n_units[1:],
                    activations
                )
            )
        )
    
    def forward(self, x: Tensor) -> Tensor:
        return self.layers(x)

    def configure_optimizers(self) -> Optimizer:
        return Adam(self.parameters(), lr=1e-4)

    def training_step(
        self, train_batch: Batch, batch_idx: int
        ) -> Tensor:
        features, labels = train_batch
        y_pred = self.forward(features)
        loss_val = cross_entropy(y_pred, labels)
        self.log("train_loss", loss_val)
        return loss_val
    
    def validation_step(
        self, val_batch: Batch, batch_idx: int
        ):
        features, labels = val_batch
        y_pred = self.forward(features)
        loss_val = cross_entropy(y_pred, labels)
        self.log("val_loss", loss_val)

logger = CSVLogger(save_dir="logs/", name="losses.csv")
trainer = Trainer(
    max_epochs=300,
    logger=logger, log_every_n_steps=10,
    callbacks=[
        EarlyStopping(monitor="val_loss", min_delta=1e-4, patience=20),
        ]
)

linear = lambda x: x
model = MLP(
    n_units=[features.shape[1], 16],
    activations=[linear],
    dropout=0.0
    )

trainer.fit(model, train_dl, val_dl)

# Evaluación

df = pd.read_csv("logs/losses.csv/version_0/metrics.csv")
df_train = (
    df
    .filter(["epoch", "train_loss"])
    .dropna()
    .groupby("epoch")
    .agg({"train_loss": "mean"})
)
df_test = (
    df
    .filter(["epoch", "val_loss"])
    .dropna()
    .groupby("epoch")
    .agg({"val_loss": "mean"})
)   

df = df_train.join(df_test, on="epoch").reset_index()

fig = (
    px.line(df, x="epoch", y=["train_loss", "val_loss"])
)
fig.show()

metric = []
for features_batch, labels_batch in test_dl:
    y_pred = torch.argmax(model(features_batch), dim=1)
    metric.append(y_pred == labels_batch)

metric = torch.concat(metric) 
print(metric.float().mean())
